correct Sale controller and add authorization to routes

+++ src/controller/chargeController.ts
@@ -2,6 +2,8 @@ import { Request, Response } from "express";
 import Charge, { ChargeTypes } from "../models/Charge";
 import { v4 as uuidv4 } from "uuid";
 
+import { AuthenticatedRequest } from "../authentication/authMiddleware";
+
 const validateChargeValue = (value: any) => {
   const isValidValue = typeof value === "number" && value > 0 && value <= 100;
 
@@ -32,9 +34,16 @@ const validateChargeType = (type: ChargeTypes) => {
   }
 };
 
-export const getAllCharges = async (_req: Request, res: Response) => {
+export const getAllCharges = async (
+  req: AuthenticatedRequest,
+  res: Response
+) => {
+  const userId = req.user?.userId;
   try {
-    const charges = await Charge.findAll();
+    if (!userId) {
+      return res.status(401).json({ message: "Usuário não autenticado" });
+    }
+    const charges = await Charge.findAll({ where: { userId } });
 
     res.json(charges);
   } catch (error) {
@@ -42,10 +51,17 @@ export const getAllCharges = async (_req: Request, res: Response) => {
   }
 };
 
-export const getChargeById = async (req: Request, res: Response) => {
+export const getChargeById = async (
+  req: AuthenticatedRequest,
+  res: Response
+) => {
   const { id } = req.params;
+  const userId = req.user?.userId;
   try {
-    const charge = await Charge.findByPk(id);
+    if (!userId) {
+      return res.status(401).json({ message: "Usuário não autenticado" });
+    }
+    const charge = await Charge.findOne({ where: { id, userId } });
     if (charge) {
       res.json(charge);
     } else {
@@ -56,7 +72,11 @@ export const getChargeById = async (req: Request, res: Response) => {
   }
 };
 
-export const createCharge = async (req: Request, res: Response) => {
+export const createCharge = async (
+  req: AuthenticatedRequest,
+  res: Response
+) => {
+  const userId = req.user?.userId;
   const { name, value, type } = req.body;
 
   try {
@@ -65,7 +85,17 @@ export const createCharge = async (req: Request, res: Response) => {
 
     await validateChargeObject(name as string, value as number);
 
-    const newCharge = await Charge.create({ id: uuidv4(), name, type, value });
+    if (!userId) {
+      return res.status(401).json({ message: "Usuário não autenticado" });
+    }
+
+    const newCharge = await Charge.create({
+      id: uuidv4(),
+      name,
+      type,
+      value,
+      userId,
+    });
 
     res.status(201).json(newCharge);
   } catch (error) {
@@ -73,9 +103,12 @@ export const createCharge = async (req: Request, res: Response) => {
   }
 };
 
-export const updateCharge = async (req: Request, res: Response) => {
+export const updateCharge = async (
+  req: AuthenticatedRequest,
+  res: Response
+) => {
   const { id } = req.params;
-
+  const userId = req.user?.userId;
   const { name, value, type } = req.body;
 
   try {
@@ -84,9 +117,13 @@ export const updateCharge = async (req: Request, res: Response) => {
 
     await validateChargeObject(name as string, value as number);
 
+    if (!userId) {
+      return res.status(401).json({ message: "Usuário não autenticado" });
+    }
+
     const [count] = await Charge.update(
       { name, type, value },
-      { where: { id } }
+      { where: { id, userId } }
     );
 
     if (count > 0) {
@@ -101,14 +138,21 @@ export const updateCharge = async (req: Request, res: Response) => {
   }
 };
 
-export const deleteCharge = async (req: Request, res: Response) => {
+export const deleteCharge = async (
+  req: AuthenticatedRequest,
+  res: Response
+) => {
   const { id } = req.params;
+  const userId = req.user?.userId;
 
   try {
-    const deletedCharge = await Charge.findByPk(id);
+    if (!userId) {
+      return res.status(401).json({ message: "Usuário não autenticado" });
+    }
+    const deletedCharge = await Charge.findOne({ where: { id, userId } });
 
     if (deletedCharge) {
-      await Charge.destroy({ where: { id } });
+      await Charge.destroy({ where: { id, userId } });
 
       res.json(deletedCharge);
     } else {
